refactor: Move allSafes computation and sort comparator into AccountsList
