+++ src/features/myAccounts/components/AccountsList/index.tsx
@@ -1,19 +1,31 @@
 import FilteredSafes from '@/features/myAccounts/components/FilteredSafes'
 import PinnedSafes from '@/features/myAccounts/components/PinnedSafes'
-import type { AllSafeItems } from '@/features/myAccounts/hooks/useAllSafesGrouped'
+import type { AllSafeItems, AllSafeItemsGrouped } from '@/features/myAccounts/hooks/useAllSafesGrouped'
 import AllSafes from '@/features/myAccounts/components/AllSafes'
+import { getComparator } from '@/features/myAccounts/utils/utils'
+import { useAppSelector } from '@/store'
+import { selectOrderByPreference } from '@/store/orderByPreferenceSlice'
+import { useMemo } from 'react'
 
 const AccountsList = ({
   searchQuery,
-  allSafes,
+  safes,
   onLinkClick,
   isSidebar,
 }: {
   searchQuery: string
-  allSafes: AllSafeItems
+  safes: AllSafeItemsGrouped
   onLinkClick?: () => void
   isSidebar: boolean
 }) => {
+  const { orderBy } = useAppSelector(selectOrderByPreference)
+  const sortComparator = getComparator(orderBy)
+
+  const allSafes = useMemo<AllSafeItems>(
+    () => [...(safes.allMultiChainSafes ?? []), ...(safes.allSingleSafes ?? [])].sort(sortComparator),
+    [safes.allMultiChainSafes, safes.allSingleSafes, sortComparator],
+  )
+
   if (searchQuery) {
     return <FilteredSafes searchQuery={searchQuery} allSafes={allSafes} onLinkClick={onLinkClick} />
   }
