+++ src/features/myAccounts/index.tsx
@@ -1,20 +1,13 @@
 import AccountsFilter from '@/features/myAccounts/components/AccountsFilter'
 import AccountsHeader from '@/features/myAccounts/components/AccountsHeader'
 import AccountsList from '@/features/myAccounts/components/AccountsList'
-import { useMemo, useState } from 'react'
+import { useState } from 'react'
 import { Box, Divider, Paper } from '@mui/material'
 import madProps from '@/utils/mad-props'
 import css from '@/features/myAccounts/styles.module.css'
 import useWallet from '@/hooks/wallets/useWallet'
-import {
-  type AllSafeItemsGrouped,
-  useAllSafesGrouped,
-  type AllSafeItems,
-} from '@/features/myAccounts/hooks/useAllSafesGrouped'
+import { type AllSafeItemsGrouped, useAllSafesGrouped } from '@/features/myAccounts/hooks/useAllSafesGrouped'
 import classNames from 'classnames'
-import { getComparator } from '@/features/myAccounts/utils/utils'
-import { selectOrderByPreference } from '@/store/orderByPreferenceSlice'
-import { useAppSelector } from '@/store'
 import useTrackSafesCount from '@/features/myAccounts/hooks/useTrackedSafesCount'
 import { DataWidget } from '@/features/myAccounts/components/DataWidget'
 
@@ -26,15 +19,8 @@ type MyAccountsProps = {
 
 const MyAccounts = ({ safes, onLinkClick, isSidebar = false }: MyAccountsProps) => {
   const wallet = useWallet()
-  const { orderBy } = useAppSelector(selectOrderByPreference)
-  const sortComparator = getComparator(orderBy)
   const [searchQuery, setSearchQuery] = useState('')
   useTrackSafesCount(safes, wallet)
-
-  const allSafes = useMemo<AllSafeItems>(
-    () => [...(safes.allMultiChainSafes ?? []), ...(safes.allSingleSafes ?? [])].sort(sortComparator),
-    [safes.allMultiChainSafes, safes.allSingleSafes, sortComparator],
-  )
 
   return (
     <Box data-testid="sidebar-safe-container" className={css.container}>
@@ -47,12 +33,7 @@ const MyAccounts = ({ safes, onLinkClick, isSidebar = false }: MyAccountsProps) 
           {isSidebar && <Divider />}
 
           <Paper className={css.safeList}>
-            <AccountsList
-              searchQuery={searchQuery}
-              allSafes={allSafes}
-              isSidebar={isSidebar}
-              onLinkClick={onLinkClick}
-            />
+            <AccountsList searchQuery={searchQuery} safes={safes} isSidebar={isSidebar} onLinkClick={onLinkClick} />
           </Paper>
         </Paper>
